+++ src/MixedModelsNLoptExt.jl
@@ -97,8 +97,7 @@ function MixedModels.optimize!(m::GeneralizedLinearMixedModel, ::NLoptBackend;
 end
 
 function NLopt.Opt(optsum::OptSummary)
-    lb = optsum.lowerbd
-    n = length(lb)
+    n = length(optsum.initial)
 
     if optsum.optimizer == :LN_NEWUOA && isone(n) # :LN_NEWUOA doesn't allow n == 1
         optsum.optimizer = :LN_BOBYQA
@@ -110,11 +109,10 @@ function NLopt.Opt(optsum::OptSummary)
     if length(optsum.xtol_abs) == n  # not true for fast=false optimization in GLMM
         NLopt.xtol_abs!(opt, optsum.xtol_abs) # absolute criterion on parameter values
     end
-    #    NLopt.lower_bounds!(opt, lb)   # use unconstrained optimization even for :LN_BOBYQA
     NLopt.maxeval!(opt, optsum.maxfeval)
     NLopt.maxtime!(opt, optsum.maxtime)
     if isempty(optsum.initial_step)
-        optsum.initial_step = NLopt.initial_step(opt, optsum.initial, similar(lb))
+        optsum.initial_step = NLopt.initial_step(opt, optsum.initial)
     else
         NLopt.initial_step!(opt, optsum.initial_step)
     end